nEW CHANGES

+++ app/Http/Controllers/ListingController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\Models\Listing;
+use App\Models\User;
 use Illuminate\Http\Request;
 use Illuminate\Validation\Rule;
 
@@ -42,8 +43,8 @@ class ListingController extends Controller
         if($request->hasFile('logo')) {
             $formFields['logo'] = $request->file('logo')->store('logos', 'public');
         }
-
-       
+    
+        $formFields['user_id'] = auth()->id();
 
         Listing::create($formFields);
 
@@ -57,6 +58,9 @@ class ListingController extends Controller
     
     // Update Listing Data
     public function update(Request $request, Listing $listing) {
+        if($listing->user_id!=auth()->id()){
+            abort(403,'Unauthorized Action');
+        }
         $formFields = $request->validate([
             'title' => 'required',
             'company' => ['required'],
@@ -80,8 +84,17 @@ class ListingController extends Controller
     
     //Delete Lisitng
     public function destroy(Listing $listing){
+        if($listing->user_id!=auth()->id()){
+            abort(403,'Unauthorized Action');
+        }
          $listing->delete();
          return redirect('/')->with('message', 'Listing deleted successfully!');
+    }
+
+    //Manage Lisitng
+    public function manage(){
+        //dd(auth()->user()->listings()->get());
+        return view('listings.manage',['listings'=>auth()->user()->listings()->get()]);
     }
