+++ routes/web.php
@@ -1,8 +1,9 @@
 <?php
 
-use App\Http\Controllers\ListingController;
-use Illuminate\Support\Facades\Route;
 use App\Models\Listing;
+use Illuminate\Support\Facades\Route;
+use App\Http\Controllers\UserController;
+use App\Http\Controllers\ListingController;
 /*
 |--------------------------------------------------------------------------
 | Web Routes
@@ -17,22 +18,42 @@ use App\Models\Listing;
 Route::get('/', [ListingController::class,'index']);
 
 //Show create form
-Route::get('/listings/create', [ListingController::class,'create']);
+Route::get('/listings/create', [ListingController::class,'create'])->middleware('auth');
 
 //Store Listing Data
-Route::post('/listings',[ListingController::class,'store']);
+Route::post('/listings',[ListingController::class,'store'])->middleware('auth');
 
 // Show Edit Form
-Route::get('/listings/{listing}/edit', [ListingController::class, 'edit']);
+Route::get('/listings/{listing}/edit', [ListingController::class, 'edit'])->middleware('auth');
 
 //Edit Submit to Update
-Route::put('/listings/{listing}',[ListingController::class,'update']);
+Route::put('/listings/{listing}',[ListingController::class,'update'])->middleware('auth');
 
 //Delete Lisitng
-Route::delete('/listings/{listing}',[ListingController::class,'destroy']);
+Route::delete('/listings/{listing}',[ListingController::class,'destroy'])->middleware('auth');
+
+//Manage Listings
+Route::get('/listings/manage',[ListingController::class,'manage'])->middleware('auth');
 
 //Single Listing
 Route::get('/listings/{listing}',[ListingController::class,'show']);
+
+//Show Register/Create Form
+Route::get('/register',[UserController::class,'create'])->middleware('guest');
+
+//Show Register/Create Form
+Route::post('/users',[UserController::class,'store']);
+
+//Log User Out
+Route::post('/logout',[UserController::class,'logout'])->middleware('auth');
+
+//Show Login Form
+Route::get('/login',[UserController::class,'login'])->name('login')->middleware('guest');
+
+//Authenticate user
+Route::post('/users/authenticate',[UserController::class,'authenticate']);
+
+
